Refactor code by moving mime into util, add TODOs for responses. Server functionality moved to server-config and server.js only start the server
